refactor weather detail scene
